requirements.txt
streamlit
django>=3.2,<4.0
celery==5.2.7
django-celery-results==2.4.0
django-celery-beat
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
setuptools
Cython
wheel
pymongo==3.12.3
redis==4.4.2
sqlparse==0.2.4
absl-py==1.4.0
accelerate==0.27.2
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
alembic==1.12.0
altair==5.1.1
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.3
attrs==19.3.0
Automat==0.8.0
autotrain-advanced==0.6.31
bitsandbytes==0.41.3
blinker==1.4
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
cmaes==0.10.0
cmake==3.27.2
codecarbon==2.2.3
colorama==0.4.3
colorlog==6.7.0
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.0
cryptography==2.8
cycler==0.11.0
datasets==2.14.5
diffusers==0.26.3
dill==0.3.7
docstring-parser==0.15
einops==0.6.1
et-xmlfile==1.1.0
evaluate==0.3.0
exceptiongroup==1.1.3
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.42.1
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.6.0
fuzzywuzzy==0.18.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
gradio==3.39.0
gradio-client==0.5.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub>=0.20.2
idna==3.4
imageio==2.31.3
importlib-metadata
importlib-resources==6.0.1
invisible-watermark==0.2.0
ipadic==1.0.0
Jinja2==3.1.2
jiwer==3.0.2
joblib==1.3.1
jsonschema==3.2.0
kiwisolver==1.4.5
lazy-loader==0.3
lit==16.0.6
loguru==0.7.0
Mako==1.2.4
Markdown==3.4.4
markdown-it-py
MarkupSafe==2.1.3
matplotlib==3.7.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
openpyxl==3.1.2
optuna==3.3.0
orjson==3.9.7
packaging==23.1
pandas==2.0.3
peft==0.9.0
Pillow==10.0.0
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydantic==1.10.11
pydub==0.25.1
Pygments==2.16.1
pynvml==11.5.0
pyparsing==3.1.1
pyrsistent==0.15.5
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
rapidfuzz==2.13.7
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.5.3
rsa==4.9
sacremoses==0.0.53
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
shtab==1.6.4
six==1.14.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.7.10
tiktoken==0.5.1
tokenizers<0.19
toolz==0.12.0
torch==2.0.1
tqdm==4.65.0
transformers<4.38
triton==2.0.0
trl==0.7.11
typing-extensions==4.7.1
tyro==0.5.11
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.23.2
websockets==11.0.3
Werkzeug==2.3.6
xgboost==1.7.6
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
google-auth-httplib2 
google-api-python-client 
oauth2client
gunicorn
psycopg2-binary>=2.9.1
